+++ NEWS.md
@@ -14,10 +14,29 @@
 
 ## New features
 
+- Our methods now can estimate the COI without grouping data points into
+  buckets (#17).
+- Within sample minor allele frequencies are now weighed by read coverage. If no
+  coverage is supplied, the coverage is assumed to be uniform across all loci
+  (#16.)
 - New `theme_coiaf()` creates a custom theme for this package.
+
+## Bug Fixes
+
+- Midpoint calculation for buckets has been improved. Previously, if buckets
+  consisted of a single value, we would combine the bucket with another bucket.
+  We no longer do so and let buckets consist of only a single value (#19).
+- In our estimation methods, each bucket is now weighed by the number of points
+  in each bucket (#15).
+- The Frequency Method is undefined for a COI less than two. When we suspect the
+  COI should be less than two, we return `NaN`. We additionally return a note
+  indicating that the COI may be one. If we are able to compute the COI, we also
+  return the COI estimated by our methods (#14, #21).
 
 ## Maintenance
 
+- PLAF has been renamed to PLMAF to indicate the focus is the minor allele.
+- WSAF has been renamed to WSMAF to indicate the focus is the minor allele.
 - Fix partial argument match warnings.
 - Update license year.
 - The number of dependencies has been reduced.